+++ lib/loan_creator.rb
@@ -8,6 +8,7 @@ module LoanCreator
   BIG_DECIMAL_DIGITS = 14
 
   autoload :ExcelFormulas,            'loan_creator/excel_formulas'
+  autoload :TimeHelper,               'loan_creator/time_helper'
   autoload :BorrowerTimetable,        'loan_creator/borrower_timetable'
   autoload :Common,                   'loan_creator/common'
   autoload :Standard,                 'loan_creator/standard'
