+++ lib/loan_creator/common.rb
@@ -1,13 +1,7 @@
 module LoanCreator
   class Common
     extend BorrowerTimetable
-
-    PERIODS_IN_MONTHS = {
-      month: 1,
-      quarter: 3,
-      semester: 6,
-      year: 12
-    }.freeze
+    include TimeHelper
 
     REQUIRED_ATTRIBUTES = [
       :period,
@@ -33,6 +27,7 @@ module LoanCreator
       realistic_durations: false
     }.freeze
 
+
     attr_reader *REQUIRED_ATTRIBUTES
     attr_reader *OPTIONAL_ATTRIBUTES.keys
 
@@ -45,15 +40,6 @@ module LoanCreator
       set_initial_values
       validate_initial_values
       prepare_custom_term_dates if term_dates?
-    end
-
-    def periodic_interests_rate(start_date, end_date)
-      if realistic_durations?
-        compute_realistic_periodic_interests_rate(start_date, end_date)
-      else
-        @periodic_interests_rate ||=
-          annual_interests_rate.div(12 / PERIODS_IN_MONTHS[period], BIG_DECIMAL_DIGITS).div(100, BIG_DECIMAL_DIGITS)
-      end
     end
 
     def timetable_term_dates
@@ -234,61 +220,6 @@ module LoanCreator
 
     def term_zero?
       (interests_start_date && interests_start_date < term_zero_date) && !term_dates?
-    end
-
-    def leap_days_count(start_date, end_date)
-      start_year = start_date.year
-      # mostly no op but allows to skip one iteration if end date is january 1st
-      end_year = (end_date - 1.day).year
-
-      (start_year..end_year).sum do |year|
-        next 0 unless Date.gregorian_leap?(year)
-
-        current_start_date =
-          if start_year == year
-            start_date
-          else
-            Date.new(year, 1, 1)
-          end
-
-        current_end_date =
-          if end_year == year
-            end_date
-          else
-            Date.new(year + 1, 1, 1)
-          end
-
-        current_end_date - current_start_date
-      end
-    end
-
-    def compute_realistic_periodic_interests_rate(start_date, end_date)
-      total_days = end_date - start_date
-      leap_days = bigd(leap_days_count(start_date, end_date))
-      non_leap_days = bigd(total_days - leap_days)
-
-      annual_interests_rate.mult(
-        leap_days.div(366, BIG_DECIMAL_DIGITS) +
-        non_leap_days.div(365, BIG_DECIMAL_DIGITS),
-        BIG_DECIMAL_DIGITS
-      ).div(100, BIG_DECIMAL_DIGITS)
-    end
-
-    # for terms spanning more than a year,
-    # we capitalize each years until the last one which behaves normally
-    def multi_part_interests(start_date, end_date)
-      duration_in_days = end_date - start_date
-      leap_days = leap_days_count(start_date, end_date)
-      non_leap_days = duration_in_days - leap_days
-
-      ratio = (non_leap_days / 365.0) + (leap_days / 366.0)
-      full_years, year_part = ratio.divmod(1)
-      rate = annual_interests_rate.div(100, BIG_DECIMAL_DIGITS)
-
-      total = amount_to_capitalize.mult((1 + rate)**full_years, BIG_DECIMAL_DIGITS)
-                                  .mult(1 + rate * year_part, BIG_DECIMAL_DIGITS)
-
-      total - amount_to_capitalize
     end
 
     def realistic_durations?
